Added functionality to allow input commands
Added a continuous loop that allows processing of certain commands.  The
commands (not all of them have been implemented yet) are LD (load), DC
(discard), SV (save), RM (remove), CP (copy), IN (insert), and EX
(exit).  To load and save a file using this new code, run the program
and type in LD, then hit enter.  Then, enter the name of the mp3 file to
input and hit enter once again.  Wait until the program reports that the
load operation is complete, then enter SV and hit enter.  Then, enter
the name of the mp3 file to write and hit enter again.  Wait until the
program reports that the save operation is complete.  Then, enter EX and
hit enter to exit the program.  The code I added serves to implement
this command line functionality.