Added copy command functionality
I added rudimentary copy command functionality for WAV files.  The copy
command takes two arguments - the second marker of the start of the
section to copy and the second marker of the end of the section to copy.
It will transfer the copied section to a separate area of memory.  (To
do so, it can cut off the end of the WAV file initially loaded.)  To
test, one can use the copy command on a WAV file and then save
immediately afterward - the copied section should play after the
original part.